Add Read-Only API Endpoints
This commit adds all 19 current API resources, but only gives access to
index and show (read) the data. The ability to create or update is
coming soon.

API Resources Added:
 - Get-ITGlueConfigurationInterfaces
 - Get-ITGlueConfigurationStatuses
 - Get-ITGlueConfigurationTypes
 - Get-ITGlueConfigurations
 - Get-ITGlueContactTypes
 - Get-ITGlueContacts
 - Get-ITGlueCountries
 - Get-ITGlueFavoriteOrganizations
 - Get-ITGlueLocations
 - Get-ITGlueManufacturers
 - Get-ITGlueModels
 - Get-ITGlueOperatingSystems
 - Get-ITGlueOrganizationStatuses
 - Get-ITGlueOrganizationTypes
 - Get-ITGlueOrganizations
 - Get-ITGluePlatforms
 - Get-ITGlueRegions
 - Get-ITGlueUserMetrics
 - Get-ITGlueUsers

—

Internal functions to specify a base URI and API Key are provided as
well. A readme file will dive into more detail on how to use these
functions and get started with the API.