Fix Encoding and Setting Export Issues
Encoding was wrong which caused issues importing the module on some Windows devices. That was fixed for all files.

Additionally, functionality to properly import and export module settings was given. This **only** works on Windows right now, as the `ConvertTo-SecureString` cmdlet is currently unavailable on Mac and Linux PowerShell core ports.